+++ src/days/1.ts
@@ -7,5 +7,7 @@ export const sunday: Day = {
 		ar: 'الأحد',
 		de: 'Sonntag',
 		sv: 'Söndag',
+		uk: 'Неділя',
+		he: 'יום ראשון',
 	},
 }