+++ src/days/4.ts
@@ -7,5 +7,7 @@ export const wednesday: Day = {
 		ar: 'الأربعاء',
 		de: 'Mittwoch',
 		sv: 'Onsdag',
+		uk: 'Середа',
+		he: 'יום רביעי',
 	},
 }