+++ src/days/5.ts
@@ -7,5 +7,7 @@ export const thursday: Day = {
 		ar: 'الخميس',
 		de: 'Donnerstag',
 		sv: 'Torsdag',
+		uk: 'Четвер',
+		he: 'יום חמישי',
 	},
 }