+++ src/days/7.ts
@@ -7,5 +7,7 @@ export const saturday: Day = {
 		ar: 'السبت',
 		de: 'Samstag',
 		sv: 'Lördag',
+		uk: 'Субота',
+		he: 'שבת',
 	},
 }